requirements.txt
ollama
openai-whisper
piper-tts
sounddevice
scipy
pyaudio